Quiz.jsで宣言した関数をhelpersに記述変更
